changed the parameters of invincible function, added spike prefab

+++ Assets/Scripts/Player/HealthBar.cs
@@ -7,14 +7,14 @@ public class HealthBar : MonoBehaviour
     //Image healthbar;  
     public Slider slider;    //the red health bar sprite
 
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
     {
         slider.maxValue = health;
         slider.value = health;
     }
 
     //value should never be set above max health
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
     }
